MAINT refactoring; remove sklearn code, part 2
