+++ skhubness/neighbors/__init__.py
@@ -9,7 +9,7 @@ from ._annoy import AnnoyTransformer, LegacyRandomProjectionTree
 from .base import VALID_METRICS, VALID_METRICS_SPARSE
 from .classification import KNeighborsClassifier, RadiusNeighborsClassifier
 from ._nmslib import NMSlibTransformer, LegacyHNSW
-from .lsh import FalconnLSH, LegacyPuffinn
+from ._puffinn import FalconnLSH, LegacyPuffinn
 from ._ngt import NGTTransformer, LegacyNNG
 from .regression import KNeighborsRegressor, RadiusNeighborsRegressor
 from .unsupervised import NearestNeighbors
