--- skhubness/neighbors/quad_tree.py
@@ -1,5 +1,0 @@
-# SPDX-License-Identifier: BSD-3-Clause
-
-from sklearn.neighbors.quad_tree import _QuadTree
-
-__all__ = ['_QuadTree']
