+++ skhubness/reduction/__init__.py
@@ -5,31 +5,31 @@
 The :mod:`skhubness.reduction` package provides methods for hubness reduction.
 """
 
-from .base import NoHubnessReduction
-from .mutual_proximity import GraphMutualProximity, MutualProximity
-from .local_scaling import GraphLocalScaling, LocalScaling
-from .dis_sim import GraphDisSimLocal, DisSimLocal
+from ._base import NoHubnessReduction
+from ._mutual_proximity import GraphMutualProximity, MutualProximity
+from ._local_scaling import GraphLocalScaling, LocalScaling
+from ._dis_sim import GraphDisSimLocal, DisSimLocal
 
 #: Supported hubness reduction algorithms
 hubness_algorithms = [
-    'mp',
-    'ls',
-    'dsl',
+    "mp",
+    "ls",
+    "dsl",
 ]
 hubness_algorithms_long = [
-    'mutual_proximity',
-    'local_scaling',
-    'dis_sim_local',
+    "mutual_proximity",
+    "local_scaling",
+    "dis_sim_local",
 ]
 
 
 __all__ = [
-    'NoHubnessReduction',
+    "NoHubnessReduction",
     "GraphLocalScaling",
-    'LocalScaling',
+    "LocalScaling",
     "GraphMutualProximity",
-    'MutualProximity',
+    "MutualProximity",
     "GraphDisSimLocal",
-    'DisSimLocal',
-    'hubness_algorithms',
+    "DisSimLocal",
+    "hubness_algorithms",
 ]
