+++ skhubness/reduction/_dis_sim.py
@@ -11,9 +11,9 @@ from sklearn.metrics import euclidean_distances
 from sklearn.utils.extmath import row_norms
 from sklearn.utils.validation import check_is_fitted, check_consistent_length, check_array
 
-from .base import GraphHubnessReduction, HubnessReduction
-from ..utils.check import check_kneighbors_graph, check_matching_n_indexed
-from ..utils.helper import k_neighbors_graph
+from ._base import GraphHubnessReduction, HubnessReduction
+from ._utils import check_kneighbors_graph, check_matching_n_indexed
+from skhubness.reduction._utils import k_neighbors_graph
 
 
 class GraphDisSimLocal(GraphHubnessReduction, TransformerMixin):
